requirements.txt
# Required Waggle Python modules
pywaggle[vision]==0.52.*

# Custom Python modules. Please add your requirements below.
numpy
